Reformat depot_tools files to 4 space indents and update .style.yapf.
The current status quo is supposedly:
- new files get 4 space indents
- existing files with 2 space indents preserve 2 space indents

In practice what happens is:
- all files get 2 space indents due to .style.yapf and the presubmit
  checks.
- added latency to CLs when landing a change is blocked on resolving
  discussions around indents

To avoid a situation where:
- contributors have to consciously ignore a presubmit check one way or
  another
- how many spaces to indent randomly shows up as a topic of contention
  in code reviews

This CL simply reformats everything to 4 spaces using Pyink and updates
the default style configuration to match.

The option of updating the presubmit checks to accomodate "legacy files
use 2 space indents, new files use 4 space indents" was also
investigated. Unfortunately, implementing this is not very practical,
since the presubmit check is based on yapf which:
- only allows static configuration of the indent level
- adding heuristics to the presubmit check to dynamically configure the
  indent level means also duplicating yapf's logic for determining style
  (https://github.com/google/yapf#formatting-style) in the presubmit
  check

Change-Id: I4369ab269e46dc60b0f744407cf73510dace5dca