Detect missing encoding parameter to open
Do not land!

This change attempts to detect missing uses of encoding='utf-8' in open
calls in PRESUBMIT.py files by injecting an override of open into the
start of the code.

The results are put in c:\src\temp\missing_encoding.txt.

When making fixes newline='' should be added when appropriate (when
opening files for writing).

Bug: 1418846
Change-Id: I1d51d88fbf6d9429f3f9daa1bc96a1618ad405bd