gclient: Add skip-git experiment.
This experiment causes all dependencies of type 'git' to be skipped for
all commands. This can be useful if you need to build Chromium in a
working tree that is guaranteed to contain a consistent, pre-synced
snapshot of the code that's managed by Git, but still need to install
cipd dependencies and run the hooks to download the remaining
dependencies.

Bug: b/310485239
Change-Id: Iee1e6a4d62b867b6d7235903817f927a9e806dc1