Fix compile_single_file with remoteexec
When using remoteexec, calling ninja directly instead of ninja_reclient
causes the build to hang. This also breaks compile_single_file since it
calls ninja directly. To fix this, switch it to call autoninja instead
which will take care of this and other details like using autosiso in
the near future.

Also, using autoninja has the side-effect of using a higher -j value
which can help when building a single file causes a lot of other files
to also be built.

Bug: 1502988
Change-Id: Ib86c5c820f7c53a8420973f1b83bf8e711a9af09