Add mac-specific reclient tweak.
Set cas_concurrency for mac builds to 100.  It doesn't affect
performance on high-speed networks, but has shown an improvement on
networks prone to congestion.

Note that these adjustments have been checked out for builds on macs, and the cas_concurrency value being a default of 100 for mac doesn't mean it should be the default for other oses.

Bug: b/304616237
Change-Id: I9431fdb35737f9f2b236296625af2f2a31e52754