cros: expand to cover more roots
The background is we maintain a fork of a fork of cros and as we're
moving to submodules this is what our repo looks like e.g. chromite is
at the top, and it's using submodules.

It doesn't seem like this would negatively affect anyone else, and would
make our system a lot simpler