Allow for more granular control over hook runs
We encountered a situation where we needed to run only one specific hook
on chromium/src (node_linux64) and all the other hooks would have wasted too much time.

With this we introduce a way to select one or more hooks to be run by name, or exclude some hooks by name. Exclusion was added only for generality completeness.

Bug: none
Change-Id: Icc32f6fbba74464bd178d57b50af51064213a2a0