Allow for more granular control over hook runs
With this we introduce a way to select one or more hooks to be run by name, or exclude some hooks by name. Exclusion was added only for generality completeness.

Inclusion and exclusion are mutually exclusive as we cannot see any scenario where using both options would make sense.

Context: We want to automate the rolling of browser_protocol from chromium/src to devtools. Our rollers can handle execution of arbitrary scripts that update the target repo files #1. The one piece that's missing it a dependency on nodejs that is handled in chromium/src via a hook. It would make little sense rearchitect the current solution (by injecting the necessary config to run that hook) if we could just reach for that hook at the exact moment we need it. Also we would like to avoid actually running all resolved hooks when we only need linux64_node.

#1 https://source.chromium.org/chromium/chromium/tools/build/+/main:recipes/recipe_modules/v8_auto_roller/script_handlers.py;l=25

Bug: chromium:1454378
Change-Id: Icc32f6fbba74464bd178d57b50af51064213a2a0