[lucicfg] Update v1.43.1 -> v1.43.2.
The most notable change is the addition of the NEWEST_FIRST
luci-scheduler policy. Note: this CL doesn't set the package's
git_revision to the exact commit where the version was updated and then
rolled into infra/infra because the package builders were down at the
time. Instead, it updates to the first good commit on the package
builders after (374d7c95).

Change-Id: Ie193d70ca41d86e549ac7026d0027528b910c9f1