Fix 'git cl format' on renamed .cc and .h files
https://chromium-review.googlesource.com/c/chromium/tools/depot_tools/+/4904205
introduced the behavior that git moved/copied files are not
automatically reformatted by git cl format. This is desired behavior if
you move a file.

If, however, you rename a class, that entails renaming a files. If the
new class name has a different length than the old class name, you end
up with not properly `git cl format`ed code.

For .cc and .h files, we can use the smarts of clang-format-diff.py.
Instead of filtering the list of changed files (removing the files which
were copied/renamed), we take a regular `git diff`, remove irrelevant
content (removing e.g. .xml files) and pass that to
clang-format-diff.py.

Bug: 737568
Change-Id: I860effcdac62c81ec68a8fa81442e1ed1ecd1ed7