Fix git cl split sending all CLs to CQ
Default the --cq-dry-run flag to false instead of None to prevent
uploaded CLs from being sent to the CQ. Log the value of cq_dry_run when
running a git cl split dry run. (It was non obvious in this case that
cq-dry-run was set to None instead of the expected value False, leading
to all uploaded CLs being sent to the CQ for a dry run.)

Additionally, improve the warning printed when more than 10 CLs will be
created to be printed whether or not the CLs are being sent to the CQ
and provide guidence on adjusting the number of CLs using the max-depth
flag. This makes the tool more user friendly by preventing a new user
from unknowingly creating many CLs without realizing there is a
supported option to customize the depth.

Change-Id: I6feba24d06dd5e88d9005345d5b0a1e15c53e197