Provide the gcloud SDK via CIPD
Siso and reclient need the gcloud SDK to get an authentication token for
RBE. Our dogfooders told us that they would prefer not to have to
install the gcloud SDK manually. Luckily, a CIPD package already exists
and we should be able to just use that for our purposes, by adding a
little "gcloud" and "gcloud.bat" wrapper to depot_tools.

Bug: b/316921158
Change-Id: I8525005d60559eadba378b689e84328ee39625b1