[scm] Implement caching for git config
This change implements a simple in-memory cache for git config calls.
The calls through this interface caches the entire config and uses it to
serve queries until it gets invalidated by a SetConfig call. This will
reduce the number of subprocess calls made by gclient.

The follow-up CLs will update the rest of the code to use this wrapper.

Bug: 1501984
Change-Id: I751a1f08eb9ae7141b8b4e1517731ed553328c03