Ignore errors on racing cleanup.
Occasionally on windows there seems to be a racing problem where rmtree
tries to delete something that no longer exists.  This seems to happen
when a build is cancelled/terminated while in-progress, so racing
actions in reproxy might have deleted the temporary file while autoninja
is cleaning up stuff.  In any case, missing a file in this case isn't
something to fail over.

Bug: b/329738642
Change-Id: I86240e31093c6415707878277b01db5931b175ba