Validate tar file links in gclient GCS
This utilizes the same tarfile validation as
download_from_google_storage, but does not retain the same 1:1 behavior
due to the `prefix` arg passed in. download_from_google_storage assumes
that the extracted content is always one directory and therefore can
check that all the links inside stay inside that one directory. However,
with some hooks that will be migrated, there may be more than one
directory. As a result, the validation will check that the links
inside the tarfile stay inside the output_dir which is one level above
the extracted directories. So if two directories are extracted, bin/
and lib/, a link inside bin/ can validly target a file inside lib/.

Bug: b/324418194
Change-Id: I73505c72084d54ca8088213dbc40b371675f8dc5