Resolve git.bat to git.exe without using shell=True
In order for git.bat to be run on Windows and ultimately run git.exe it
is normally necessary to pass shell=True to Popen. However this means
invoking cmd.exe every time we want to invoke git.exe, which ends up
being (as of today) over 1,500 invocations of cmd.exe when running the
"update" stage of gclient sync. This makes the "update" stage take about
1.45x as long to run - it's a significant overhead.

This change teaches subprocess2.py to special-case invoking 'git' and
parse the .exe name out of the .bat file, for some lovely speedups.

Bug: 332982922
Change-Id: I4e052444276d55c0496144b9816fe365a95a56d6