Match patch behavior for bot_update test data to production.
In production, if there is no patch, then the bot_update output will
have the value for the patch_root key set to null and the patch_failure
key will only be set if there was a patch failure. This updates the
output_json test API method so that the json matches this behavior,
changing the root parameter to patch_root since that's the only use of
the parameter.

Change-Id: I0c34a41ff18bce2900a2f176ff992e6d147abbd4