Add return_custom_result flag to bot_update.ensure_checkout.
In order to facilitate removing uses of api.path.checkout_dir from
downstream repos, this change adds a custom return type for
bot_update.ensure_checkout that can be enabled with the
return_custom_result argument. If set to True, a Result object will be
returned that records the relevant paths (directory where the checkout
was performed, the repo that was checked out and the repo that was
patched, if any). This provides the caller the ability to work in any of
these directories without using api.path.checkout_dir and without
requiring boilerplate to construct the paths. It also includes some
attributes that provide details from within the json output to abstract
that out.

Bug: 329113288, 339472834
Change-Id: I2ec6db635c5b799bdb65d4e9364e7d99aae4159e