[depot_tools] auto correct url always
[1] was originally added by an [ancient CL] in 2013.
It runs `git config ...` and checks if the auto-correct the URL
if the output is not "False".

It has two problems.
1) `git config` never returns False. It may return false, not False.
   from https://git-scm.com/docs/git-config
   bool: canonicalize values as either "true" or "false".

2) the inline comment, describing the intention, is confusing.
   "Skip url auto-correction if remote.origin.gclient-auto-fix-url is set"

   It's not clear whether it's set to any value? false? true?
   Nevertheless, what the code actually does is to auto-correct the URL,
   unless the config is set to False, which can never be true, as
   described in (1).

As a result, the url was auto-corrected, howver
"remote.origin.gclient-auto-fix-url" was set or not.
That's how the code had behaved from 2013 to 2024.

In Feb 2024, https://crrev.com/c/5280779 changed the behaviour.
It replaces the dubious check with GetConfigBool(), which may sound
correct by itself.
i.e., auto-correct if "remote.origin.gclient-auto-fix-url" is set to
"true".

However, it doesn't match the existing behaviour, which
auto-correct the URL regardless of the config. Depot tools had the
behaviour for 11 years. Fixing it is probably not the best decision.

[1]: https://source.chromium.org/chromium/chromium/tools/depot_tools/+/3b9212b7ee01b13d2f8fb9a41bc2639fd391df5c:gclient_scm.py;l=764-768
[ancient CL]: https://chromiumcodereview.appspot.com/15325003

Bug: 342445995
Change-Id: Idc62f803f2ef1beb11a93a58432867ce16226da7