[gerrit_util] Add opt-in SSOAuthenticator.
This adds a new authenticator implementation to gerrit_util which
uses the Google corp `git-remote-sso` protocol helper to get
secure tunneled access to the internal hosted Google git services.

It is possible to opt in on Google-managed devices by adding the
following git configuration:

    [url "sso://chromium.git.corp.google.com/"]
        insteadOf = https://chromium.googlesource.com/
        insteadOf = http://chromium.googlesource.com/
    [depot-tools]
      use-new-auth-stack = 1

This CL was uploaded using the new authenticator.

R=ayatane@google.com, ddoman@chromium.org, yiwzhang@google.com

Bug: 336351842
Change-Id: Ia3eccb77ebf1cdc863513ac095bf65ede6adda43