remove weekly and wtf commands
Those 2 commands have been broken since the introduction of gcs dep as
it tries to run a git command aginst path to a gcs dep. Since nobody
has complained so far, it's safe to assume the command is never used
by anyone today, therefore, it should be safe to delete.

Bug: 345486495
Change-Id: Ia27e5e8486f790234bd66e890e995f1d5c14f9c5