Fix git diff-dependent "cl" commands that may create very long command lines.
Running a command line longer that 32k on Windows lead to issues (as usual).
We run git duff in multiple scenarios on a list of changed
files. This list is not partitioned and may lead to a command line much
longer that we can allow.

The change replaces BuildGitDiffCmd with RunGitDiffCmd that splits the
file list internally and runs git diff multiple times if required.

Change-Id: Icfd8efc7520105b8d98125e90bb96c2bb9bd3a0a