Fix git diff calls when a lot of files modified in a branch
Running a command line longer that 32k on Windows lead to issues (as usual). We run git diff in multiple scenarios on a list of changed
files. This list is not partitioned and may lead to a command line much
longer than OS can allow.

This change replaces BuildGitDiffCmd with RunGitDiffCmd that splits the
file list internally and runs git diff multiple times if required.

Change-Id: Icfd8efc7520105b8d98125e90bb96c2bb9bd3a0a