honor the cl description from the option even when CL is uploaded
Before, if a CL has already been uploaded for a local change and then
the developer run `git cl upload --message="xxx"`, the new message will
be silently ignored and the description of the existing CL will be used.

The new behavior can also be used as a workaround to skip banned word
check when there is a banned word in the CL description. `git cl upload
--edit-description` will open a live editor which is hard to work with
in a script environment and `git cl description` directly calls Gerrit
REST api and the Gerrit REST api doesn't have the option to skip banned
word check.

Bug: 346174899
Change-Id: I6d7e361b7da0fd06675c04df4fd2f68c7954dd40