Create scm.DIFF.GetAllFiles
For a ProvidedDiffChange, the AllFiles method naively returns all files
with rglob("*"). The returned list includes files in nested submodules.
This does not match the behavior of GitChange's AllFiles which uses
git ls-files to find all files.

Implement a new SCM that stops iterating recursively when it encounters
.gitmodules or .gitignore. This is behavior is more aligned with git
and improves performance when using --all_files with diff.

Bug: b/323243527
Change-Id: I170d0f1bc4a838acea04779dee3df7fca0bce359