[git_cl] Move cwd param to apply method
This feels like bikeshedding, but semantically, we are applying the
changes to a specific repo, and you could take one
GitAuthConfigChanger and apply it to multiple different repos.

(The config host, etc. are usually tied to a specific repo, so there
needs to be some further refactoring to make this actually decoupled
from the working directory, but keeping this small for sane code
review)

Bug: b/351071334
Change-Id: I756f1134be516e19e26e35ffca7f7ceb389aaa30