[git_cl] Add cwd param to infer_and_create
The settings should be inferred from a specific Git repo instead of
assuming the working dir.  Also this makes this slightly easier to
test.

Just adding the param for now, to focus on updating call sites, etc.

(Also, actually inferring from any arbitrary repo might be a little
tricky, since the functions we're using don't allow for that.  So
deferring that for later CL.)

Bug: b/351071334
Change-Id: I5baa84ada26fef20a2f3f5857274972da4454d67