[depot_tools] retry git config if it lock-fails
If git config set is executed, it writes the new content into
a temp file (but called lock-file) and then replace .gitconfig
with it.

However, if it cannot create the lock file, it returns an error.
With this CL, if git config fails with the lock file error,
depot_tools will retry it at most 5 times with 0.2s interval.

In crbug/351950514, the engineer reported that the lock failure happens
if they run git rebase-update in multiple checkouts in parallel.
However,
- The config set should be as quick as 0.1s.
- git rebase-update wouldn't call git config set crazy

This CL just hard-codes the retry count and interval, because
executing git rebase-update multiple times in parallel would probably
a very rare case, and 5 retries with 0.2s interval would probably be
good enough for 2 or 3 rebase-update executions in parallel.

If the lock failure happens frequently even with the retries, we may
consider adding an option for the retry count, but it'd be a bit
surprise.

Bug: 351950514
Change-Id: I985af0d8b7458dbf47cd6baa857dc5adccf15031