[depot_tools] retry git config if it lock-fails
If git config set is executed, it writes the new content into
a temp file (but called lock-file) and then replace .gitconfig
with it.

However, if it cannot create the lock file, it returns an error.
With this CL, if git config fails with the lock file error,
depot_tools will retry it at most 5 times with 0.2s interval.

In case multiple git rebase-updates are running in parallel, the lock
failure may happen and the git rebase-updates fail.

Bug: 351950514
Change-Id: I985af0d8b7458dbf47cd6baa857dc5adccf15031